Refactor data_access into a class

+++ tests.py
@@ -2,7 +2,7 @@ import os
 import sqlite3
 import unittest
 
-import data_access
+from data_access import DataAccess
 
 test_db = 'test.db'
 
@@ -18,15 +18,26 @@ def _table_exists():
 class DataAccessTestCase(unittest.TestCase):
 
     def tearDown(self):
-        os.remove(test_db)
+        if os.path.exists(test_db):
+            os.remove(test_db)
+
+    def test_constructor_with_default_params(self):
+        da = DataAccess()
+        self.assertEqual(da.database, 'expenses.db')
+        self.assertEqual(da.table_name, 'expenses')
+        da.close()
 
     def test_create_when_table_does_not_exist(self):
-        data_access.create()
+        da = DataAccess(database=test_db)
+        da.create()
+        da.close()
         self.assertEqual(_table_exists(), 1)
     
     def test_create_when_table_already_exists(self):
-        data_access.create()
-        data_access.create()
+        da = DataAccess(database=test_db)
+        da.create()
+        da.create()
+        da.close()
         self.assertEqual(_table_exists(), 1)
 
 if __name__ == '__main__':
